Move domain creator time dep opts to new directory
Also renames the files now that they are in a directory:

BinaryCompactObjectHelpers -> BinaryCompactObject
ShapeMapOptions            -> ShapeMap
SphereTimeDependentMaps    -> Sphere

Also moved create_grid_anchors out of the BCOHelpers and just
into the BCO file.

This is all in anticipation of adding more time dep opts for
different maps and not wanting to clutter the Domain/Creators
directory.
