+++ src/Domain/Creators/TimeDependentOptions/CMakeLists.txt
@@ -1,0 +1,19 @@
+# Distributed under the MIT License.
+# See LICENSE.txt for details.
+
+spectre_target_sources(
+  ${LIBRARY}
+  PRIVATE
+  BinaryCompactObject.cpp
+  ShapeMap.cpp
+  Sphere.cpp
+)
+
+spectre_target_headers(
+  ${LIBRARY}
+  INCLUDE_DIRECTORY ${CMAKE_SOURCE_DIR}/src
+  HEADERS
+  BinaryCompactObject.hpp
+  ShapeMap.hpp
+  Sphere.hpp
+  )
