+++ tests/Unit/Domain/Creators/CMakeLists.txt
@@ -1,15 +1,11 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
-add_subdirectory(Python)
-add_subdirectory(TimeDependence)
-
 set(LIBRARY "Test_DomainCreators")
 
 set(LIBRARY_SOURCES
   Test_AlignedLattice.cpp
   Test_BinaryCompactObject.cpp
-  Test_BinaryCompactObjectHelpers.cpp
   Test_BlockGroups.cpp
   Test_Brick.cpp
   Test_Cylinder.cpp
@@ -23,10 +19,10 @@ set(LIBRARY_SOURCES
   Test_RotatedBricks.cpp
   Test_RotatedIntervals.cpp
   Test_RotatedRectangles.cpp
-  Test_ShapeMapOptions.cpp
   Test_Sphere.cpp
-  Test_SphereTimeDependentMaps.cpp
   )
+
+add_subdirectory(TimeDependentOptions)
 
 add_test_library(${LIBRARY} "${LIBRARY_SOURCES}")
 
@@ -42,3 +38,6 @@ target_link_libraries(
   SphericalHarmonics
   SphericalHarmonicsIO
   )
+
+add_subdirectory(Python)
+add_subdirectory(TimeDependence)
