+++ tests/Unit/Domain/Creators/TimeDependentOptions/CMakeLists.txt
@@ -1,0 +1,9 @@
+# Distributed under the MIT License.
+# See LICENSE.txt for details.
+
+set(LIBRARY_SOURCES
+  ${LIBRARY_SOURCES}
+  TimeDependentOptions/Test_BinaryCompactObject.cpp
+  TimeDependentOptions/Test_ShapeMap.cpp
+  TimeDependentOptions/Test_Sphere.cpp
+  PARENT_SCOPE)
